+++ client/python/cryoet_data_portal/src/cryoet_data_portal/_client.py
@@ -8,12 +8,14 @@ from gql.transport.requests import RequestsHTTPTransport
 
 from cryoet_data_portal._constants import USER_AGENT
 
+DEFAULT_URL = "https://graphql.cryoetdataportal.cziscience.com/v1/graphql"
+
 
 class Client:
     """A GraphQL Client library that can traverse all the metadata in the CryoET Data Portal
 
     Args:
-        url (Optional[str]): The API URL to connect to, defaults to "https://graphql.cryoetdataportal.cziscience.com/v1/graphql"
+        url (Optional[str]): The API URL to connect to, defaults to the latest portal endpoint.
 
     Returns:
         A GraphQL API Client library
@@ -25,9 +27,8 @@ class Client:
     """
 
     def __init__(self, url: Optional[str] = None):
-        # Use our default API URL
         if not url:
-            url = "https://graphql.cryoetdataportal.cziscience.com/v1/graphql"
+            url = DEFAULT_URL
         transport = RequestsHTTPTransport(
             url=url,
             retries=3,