+++ frontend/packages/data-portal/app/components/DepositionFilter/DepositionIdFilter.tsx
@@ -1,4 +1,4 @@
-import { IDFilter } from 'app/components/Filters'
+import { IdFilter } from 'app/components/Filters'
 import { IdPrefix } from 'app/constants/idPrefixes'
 import { QueryParams } from 'app/constants/query'
 import { useI18n } from 'app/hooks/useI18n'
@@ -7,7 +7,7 @@ export function DepositionIdFilter() {
   const { t } = useI18n()
 
   return (
-    <IDFilter
+    <IdFilter
       id="deposition-id-input"
       title={t('filterByDepositionId')}
       label={t('depositionId')}