+++ frontend/packages/data-portal/app/constants/annotationObjectIdLinks.ts
@@ -1,0 +1,5 @@
+export const GO = 'https://amigo.geneontology.org/amigo/term/'
+export const UNIPROTKB = 'https://www.uniprot.org/uniprotkb/'
+
+export const GO_PREFIX = 'GO:'
+export const UNIPROTKB_PREFIX = 'UniProtKB:'