+++ frontend/packages/data-portal/app/constants/idPrefixes.ts
@@ -1,4 +1,3 @@
 export enum IdPrefix {
-  Deposition = 'CZCDP-',
-  GeneOntology = 'GO:',
+  Deposition = 'CZCDP',
 }